Changed ssh to https for eglot, and from ccls to clangd
In my current project, ccls just crashes after some pretty simple
navigation, when eglot is telling it that some file is being open.

This is simply to usable. It does not seem that I have the same
problem with clangd 8. But with eglot, it won't find references.

I guess I'll give a try to `lsp-mode`.